Add GetRootSpanTool with shared types and delegation pattern
- Implement GetRootSpanTool that retrieves root span (parent_id is None)
- Create shared types.py with SpanResult to avoid type duplication
- Delegate GetRootSpanTool.invoke() to GetSpanTool for DRY principle
- Add GET_ROOT_SPAN constant to ToolNames
- Update imports and exports in __init__.py
- Add comprehensive test suite with 8 test cases
- Support pagination, attribute filtering, and error handling
