+++ mlflow/genai/judges/tools/constants.py
@@ -14,4 +14,5 @@ class ToolNames:
     """Registry of judge tool names."""
 
     GET_TRACE_INFO = "get_trace_info"
+    GET_ROOT_SPAN = "get_root_span"
     GET_SPAN = "get_span"
