+++ mlflow/genai/judges/tools/get_root_span.py
@@ -5,26 +5,13 @@ This module provides a tool for retrieving the root span of a trace,
 which contains the top-level inputs and outputs.
 """
 
-import json
-from dataclasses import dataclass
-
 from mlflow.entities.trace import Trace
 from mlflow.genai.judges.tools.base import JudgeTool
-from mlflow.genai.judges.tools.utils import create_page_token, parse_page_token
+from mlflow.genai.judges.tools.constants import ToolNames
+from mlflow.genai.judges.tools.get_span import GetSpanTool
+from mlflow.genai.judges.tools.types import SpanResult
 from mlflow.types.llm import FunctionToolDefinition, ToolDefinition, ToolParamsSchema
 from mlflow.utils.annotations import experimental
-
-
-@experimental(version="3.4.0")
-@dataclass
-class RootSpanResult:
-    """Result from getting the root span."""
-
-    span_id: str | None
-    content: str | None
-    content_size_bytes: int
-    page_token: str | None = None
-    error: str | None = None
 
 
 @experimental(version="3.4.0")
@@ -37,12 +24,12 @@ class GetRootSpanTool(JudgeTool):
 
     @property
     def name(self) -> str:
-        return "get_root_span"
+        return ToolNames.GET_ROOT_SPAN
 
     def get_definition(self) -> ToolDefinition:
         return ToolDefinition(
             function=FunctionToolDefinition(
-                name="get_root_span",
+                name=ToolNames.GET_ROOT_SPAN,
                 description=(
                     "Retrieve the root span of the trace, which contains the top-level inputs "
                     "to the agent and final outputs. Note that in some traces, the root span "
@@ -86,7 +73,7 @@ class GetRootSpanTool(JudgeTool):
         attributes_to_fetch: list[str] | None = None,
         max_content_length: int = 100000,
         page_token: str | None = None,
-    ) -> RootSpanResult:
+    ) -> SpanResult:
         """
         Get the root span from the trace.
 
@@ -97,47 +84,28 @@ class GetRootSpanTool(JudgeTool):
             page_token: Token to retrieve the next page (offset in bytes)
 
         Returns:
-            RootSpanResult with the root span content as JSON string
+            SpanResult with the root span content as JSON string
         """
         if not trace or not trace.data or not trace.data.spans:
-            return RootSpanResult(
+            return SpanResult(
                 span_id=None, content=None, content_size_bytes=0, error="Trace has no spans"
             )
 
-        root_span = None
+        root_span_id = None
         for span in trace.data.spans:
             if span.parent_id is None:
-                root_span = span
+                root_span_id = span.span_id
                 break
 
-        if not root_span:
-            return RootSpanResult(
+        if not root_span_id:
+            return SpanResult(
                 span_id=None,
                 content=None,
                 content_size_bytes=0,
                 error="No root span found in trace",
             )
 
-        offset = parse_page_token(page_token)
-        span_dict = root_span.to_dict()
-
-        if attributes_to_fetch is not None and span_dict.get("attributes"):
-            filtered_attributes = {}
-            for attr in attributes_to_fetch:
-                if attr in span_dict["attributes"]:
-                    filtered_attributes[attr] = span_dict["attributes"][attr]
-            span_dict["attributes"] = filtered_attributes
-
-        full_content = json.dumps(span_dict, default=str, indent=2)
-        total_size = len(full_content.encode("utf-8"))
-        end_offset = min(offset + max_content_length, total_size)
-        content_chunk = full_content[offset:end_offset]
-        next_page_token = create_page_token(end_offset) if end_offset < total_size else None
-
-        return RootSpanResult(
-            span_id=root_span.span_id,
-            content=content_chunk,
-            content_size_bytes=len(content_chunk.encode("utf-8")),
-            page_token=next_page_token,
-            error=None,
+        get_span_tool = GetSpanTool()
+        return get_span_tool.invoke(
+            trace, root_span_id, attributes_to_fetch, max_content_length, page_token
         )
